# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Glazz API
  version: 0.2.0
  summary: HTTP contract for the Glazz MVP
  description: |
    Source of truth for Glazz HTTP resources. Chat generation is commanded over
    the versioned WebSocket protocol. Provider-specific concepts are private.
  contact:
    name: Glazz Engineering
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: https://api.glazz.hlab.sh/api/v1
    description: Production placeholder
...
security:
  - accessCookie: []
paths:
...
  /models:
    get:
      tags: [Runtime]
      summary: List models visible to the actor
      operationId: listVisibleModels
      security:
        - accessCookie: []
        - guestCookie: []
      responses:
        "200":
          description: Models enabled for the current actor.
          content:
            application/json:
              schema:
                type: object
                required: [items, defaultModelId]
                properties:
                  items:
                    type: array
                    items: {$ref: "#/components/schemas/Model"}
                  defaultModelId: {$ref: "#/components/schemas/Id"}
        "401": {$ref: "#/components/responses/Unauthorized"}
  /guest-sessions:
    post:
      tags: [Guests]
      summary: Create or resume a guest session
      operationId: createOrResumeGuestSession
      security: []
      responses:
        "200":
          description: Existing guest session resumed.
          headers:
            Set-Cookie: {$ref: "#/components/headers/SetGuestCookie"}
          content:
            application/json:
              schema: {$ref: "#/components/schemas/GuestAllowance"}
        "201":
          description: Guest session created.
          headers:
            Set-Cookie: {$ref: "#/components/headers/SetGuestCookie"}
          content:
            application/json:
              schema: {$ref: "#/components/schemas/GuestAllowance"}
        "503": {$ref: "#/components/responses/ServiceUnavailable"}
  /guest-sessions/current:
    get:
      tags: [Guests]
      summary: Get the current guest allowance
      operationId: getCurrentGuestSession
      security: [{guestCookie: []}]
      responses:
        "200":
          description: Remaining allowance. No raw guest identifier is returned.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/GuestAllowance"}
        "401": {$ref: "#/components/responses/Unauthorized"}
  /auth/google/start:
    get:
      tags: [Authentication]
      summary: Start Google login
      operationId: startGoogleLogin
      security: []
      parameters:
        - name: returnTo
          in: query
          description: Allowlisted relative application path.
          schema: {type: string, pattern: "^/[A-Za-z0-9/_?=&.-]*$"}
        - name: termsAccepted
          in: query
          required: true
          description: Explicit acceptance of the currently advertised terms version.
          schema: {type: boolean, const: true}
        - name: privacyAccepted
          in: query
          required: true
          description: Explicit acceptance of the currently advertised privacy version.
          schema: {type: boolean, const: true}
      responses:
        "302":
          description: Redirect to Google with server-held state, PKCE, and nonce.
          headers:
            Location: {schema: {type: string, format: uri}}
        "400": {$ref: "#/components/responses/BadRequest"}
  /auth/google/callback:
    get:
      tags: [Authentication]
      summary: Complete Google login
      operationId: completeGoogleLogin
      security: []
      parameters:
        - {name: code, in: query, required: true, schema: {type: string, minLength: 1}}
        - {name: state, in: query, required: true, schema: {type: string, minLength: 20}}
      responses:
        "302":
          description: Login completed and guest conversation migrated atomically.
          headers:
            Location: {schema: {type: string, format: uri-reference}}
            Set-Cookie:
              description: Sets access and refresh HttpOnly cookies.
              schema: {type: string}
        "400": {$ref: "#/components/responses/BadRequest"}
        "409":
          $ref: "#/components/responses/Conflict"
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Rotate the refresh token
      operationId: refreshAuthSession
      security: [{refreshCookie: [], csrfHeader: []}]
      responses:
        "204":
          description: Refresh token rotated and access cookie issued.
          headers:
            Set-Cookie: {schema: {type: string}}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "409":
          description: Refresh-token reuse detected; token family revoked.
          content:
            application/problem+json:
              schema: {$ref: "#/components/schemas/ErrorEnvelope"}
              example:
                error: {code: refresh_token_reused, message: Session revoked., requestId: req_01}
  /auth/logout:
    post:
      tags: [Authentication]
      summary: Log out the current session
      operationId: logout
      security: [{accessCookie: [], csrfHeader: []}]
      responses:
        "204": {description: Current auth session revoked and cookies cleared.}
        "401": {$ref: "#/components/responses/Unauthorized"}
  /auth/ws-ticket:
    post:
      tags: [Authentication]
      summary: Create a WebSocket ticket
      operationId: createWebSocketTicket
      security:
        - accessCookie: [ ]
          csrfHeader: [ ]
        - guestCookie: [ ]
          csrfHeader: [ ]
      responses:
        "201":
          description: Single-use ticket valid for at most 30 seconds.
          content:
            application/json:
              schema:
                type: object
                required: [ticket, expiresAt]
                properties:
                  ticket: {type: string, minLength: 32}
                  expiresAt: {type: string, format: date-time}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "503": {$ref: "#/components/responses/ServiceUnavailable"}
  /me:
    get:
      tags: [Authentication]
      summary: Get the current user
      operationId: getCurrentUser
      responses:
        "200":
          description: Current user and effective permissions.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/CurrentUser"}
        "401": {$ref: "#/components/responses/Unauthorized"}
    patch:
      tags: [Authentication]
      summary: Update current-user preferences
      operationId: updateCurrentUser
      security: [{accessCookie: [], csrfHeader: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [locale]
              properties:
                locale: {type: string, enum: [en, es]}
      responses:
        "200":
          description: Updated current user.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/CurrentUser"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
    delete:
      tags: [Authentication]
      summary: Request account deletion
      operationId: requestAccountDeletion
      security: [{accessCookie: [], csrfHeader: [], recentAuth: []}]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [confirmation]
              properties:
                confirmation: {type: string, const: DELETE}
      responses:
        "202":
          description: Sessions revoked immediately; purge due within 24 hours.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/DeletionJob"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "428":
          $ref: "#/components/responses/RecentAuthRequired"
  /me/sessions:
    get:
      tags: [Authentication]
      summary: List device sessions
      operationId: listAuthSessions
      responses:
        "200":
          description: Active device sessions.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: {$ref: "#/components/schemas/AuthSession"}
        "401": {$ref: "#/components/responses/Unauthorized"}
  /me/sessions/{sessionId}:
    delete:
      tags: [Authentication]
      summary: Revoke a device session
      operationId: revokeAuthSession
      security: [{accessCookie: [], csrfHeader: []}]
      parameters:
        - {$ref: "#/components/parameters/SessionId"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      responses:
        "204": {description: Session revoked.}
        "404": {$ref: "#/components/responses/NotFound"}
  /me/reauthenticate:
    post:
      tags: [Authentication]
      summary: Start recent reauthentication
      operationId: startReauthentication
      security: [{accessCookie: [], csrfHeader: []}]
      responses:
        "200":
          description: Returns a fixed Google reauthentication URL.
          content:
            application/json:
              schema:
                type: object
                required: [authorizationUrl]
                properties:
                  authorizationUrl: {type: string, format: uri}
  /me/deletion:
    get:
      tags: [Authentication]
      summary: Get account deletion state
      operationId: getAccountDeletion
      responses:
        "200":
          description: Current deletion job.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/DeletionJob"}
        "404": {$ref: "#/components/responses/NotFound"}
  /conversations:
    get:
      tags: [Conversations]
      summary: List owned conversations
      operationId: listConversations
      security:
        - accessCookie: []
        - guestCookie: []
      parameters:
        - {$ref: "#/components/parameters/Limit"}
        - {$ref: "#/components/parameters/After"}
        - {name: query, in: query, schema: {type: string, maxLength: 200}}
        - {name: archived, in: query, schema: {type: boolean, default: false}}
      responses:
        "200":
          description: Owned conversations ordered by update time and ID.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ConversationPage"}
        "401": {$ref: "#/components/responses/Unauthorized"}
    post:
      tags: [Conversations]
      summary: Create a conversation
      operationId: createConversation
      security:
        - accessCookie: []
          csrfHeader: []
        - guestCookie: []
          csrfHeader: []
      parameters:
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody:
        required: false
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CreateConversation"}
      responses:
        "201":
          description: Conversation created. Guests may own only one.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Conversation"}
        "409": {$ref: "#/components/responses/Conflict"}
  /conversations/{conversationId}:
    parameters:
      - {$ref: "#/components/parameters/ConversationId"}
    get:
      tags: [Conversations]
      summary: Get an owned conversation
      operationId: getConversation
      security:
        - accessCookie: []
        - guestCookie: []
      responses:
        "200":
          description: Owned conversation.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Conversation"}
        "404": {$ref: "#/components/responses/NotFound"}
    patch:
      tags: [Conversations]
      summary: Update a conversation
      operationId: updateConversation
      security:
        - accessCookie: []
          csrfHeader: []
        - guestCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/UpdateConversation"}
      responses:
        "200":
          description: Updated conversation; model changes require idle state.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Conversation"}
        "409": {$ref: "#/components/responses/Conflict"}
    delete:
      tags: [Conversations]
      summary: Delete a conversation
      operationId: deleteConversation
      security:
        - accessCookie: []
          csrfHeader: []
        - guestCookie: []
          csrfHeader: []
      parameters:
        - {$ref: "#/components/parameters/IdempotencyKey"}
      responses:
        "204": {description: Conversation deleted.}
        "404": {$ref: "#/components/responses/NotFound"}
  /conversations/{conversationId}/messages:
    get:
      tags: [Conversations]
      summary: List conversation messages
      operationId: listMessages
      security:
        - accessCookie: []
        - guestCookie: []
      parameters:
        - {$ref: "#/components/parameters/ConversationId"}
        - {$ref: "#/components/parameters/Limit"}
        - {$ref: "#/components/parameters/After"}
      responses:
        "200":
          description: Durable transcript; partial terminal status is explicit.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/MessagePage"}
        "404": {$ref: "#/components/responses/NotFound"}
  /conversations/{conversationId}/retry:
    post:
      tags: [Conversations]
      summary: Retry the latest failed generation
      operationId: retryLatestGeneration
      security:
        - accessCookie: []
          csrfHeader: []
        - guestCookie: []
          csrfHeader: []
      parameters:
        - {$ref: "#/components/parameters/ConversationId"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      responses:
        "202":
          description: Latest failed or cancelled generation accepted for retry.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Generation"}
        "409": {$ref: "#/components/responses/Conflict"}
        "429": {$ref: "#/components/responses/QuotaExceeded"}
  /usage:
    get:
      tags: [Usage]
      summary: Get current quota usage
      operationId: getUsage
      security:
        - accessCookie: []
        - guestCookie: []
      responses:
        "200":
          description: Effective actor limits and current consumption.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Usage"}
...
components:
  securitySchemes:
    accessCookie:
      type: apiKey
      in: cookie
      name: glazz_access
    refreshCookie:
      type: apiKey
      in: cookie
      name: glazz_refresh
    guestCookie:
      type: apiKey
      in: cookie
      name: glazz_guest
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
    recentAuth:
      type: apiKey
      in: cookie
      name: glazz_recent_auth
  headers:
    SetGuestCookie:
      description: Signed HttpOnly guest cookie; Secure outside local development.
      schema: {type: string}
...
  responses:
...
    Unauthorized:
      description: Missing, expired, or invalid actor credentials.
      content: {application/problem+json: {schema: {$ref: "#/components/schemas/ErrorEnvelope"}}}
...
    ServiceUnavailable:
      description: Dependency, maintenance, or provider budget prevents operation.
      content: {application/problem+json: {schema: {$ref: "#/components/schemas/ErrorEnvelope"}}}
  schemas:
    Id:
      type: string
      pattern: "^[0-9a-fA-F-]{36}$"
...
    Model:
      type: object
      required: [id, name, capabilities]
      properties:
        id: {$ref: "#/components/schemas/Id"}
        name: {type: string}
        description: {type: string}
        capabilities:
          type: object
          required: [chatCompletions, markdown, code]
          properties:
            chatCompletions: {type: boolean, const: true}
            markdown: {type: boolean}
            code: {type: boolean}
...